Extract LayerSkyGridModule
