+++ src/app/app.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { AstrocadreComponent } from 'src/app/components/astrocadre/astrocadre.component';
 import { CoreModule } from '#core/core.module';
 import { ControlsModule } from '#controls/controls.module';
+import { LayerSkyGridModule } from '#layer-sky-grid/layer-sky-grid.module';
 
 /**
  * The root module of the application.
@@ -15,7 +16,8 @@ import { ControlsModule } from '#controls/controls.module';
   ],
   imports: [
     CoreModule,
-    ControlsModule
+    ControlsModule,
+    LayerSkyGridModule // TODO remove when loaded dynamically
   ],
   bootstrap: [AstrocadreComponent]
 })
