+++ src/app/modules/core/models/event/layer-hidden-event.spec.ts
@@ -1,6 +1,6 @@
 import { LayerHiddenEvent } from '#core/models/event/layer-hidden-event';
 import { AxialCurvesFactory } from '#core/models/layers/factories/axial-curves-factory';
-import { SkyGridLayerFactory } from '#core/models/layers/factories/sky-grid-layer-factory';
+import { SkyGridLayerFactory } from '#layer-sky-grid/models/sky-grid-layer-factory';
 import { mockedLayers } from '#core/test-utils/mocked-layers.spec';
 
 describe('LayerHiddenEvent', () => {
