+++ src/app/modules/layer-sky-grid/layer-sky-grid.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CoreModule } from '#core/core.module';
+
+@NgModule({
+  imports: [
+    CoreModule
+  ]
+})
+export class LayerSkyGridModule {
+
+}
