+++ src/app/modules/layer-sky-grid/models/sky-grid-layer-factory.ts
@@ -1,8 +1,8 @@
 import { LineSegments } from 'three';
+import { SkyGrid } from '#layer-sky-grid/models/sky-grid';
 import { Layer } from '#core/models/layers/layer';
 import { AxialCurvesFactory } from '#core/models/layers/factories/axial-curves-factory';
 import { LayerFactory } from '#core/models/layers/factories/layer-factory';
-import { SkyGrid } from '#core/models/layers/sky-grid';
 
 /**
  * Factory for the renderable layer of the celestial coordinates grid.
