+++ src/app/modules/layer-sky-grid/models/sky-grid.ts
